Move most descriptor-related code into descr.cpp
I think it now contains the same types that CPython's
descrobject.c handles.
